requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
faster-whisper==0.9.0
numpy==1.24.3
pydub==0.25.1
python-magic-bin==0.4.14; platform_system=="Windows"
python-magic==0.4.27; platform_system!="Windows"
python-dotenv==1.0.0
onnxruntime==1.15.1 